backend/requirements.txt
amqp
asgiref
asttokens
billiard
blinker
Bootstrap-Flask
celery
certifi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
colorgram.py
decorator
Django
django-extensions
django-filter
django-redis
djangorestframework
executing
Flask
flask-bootstrap5
Flask-Login
Flask-WTF
idna
ipython
ipython_pygments_lexers
itsdangerous
jedi
Jinja2
kombu
Markdown
MarkupSafe
matplotlib-inline
packaging
parso
pillow
prompt_toolkit
psycopg2-binary
pure_eval
Pygments
python-dateutil
razorpay
redis
requests
si
sqlparse
stack-data
traitlets
tzdata
tzlocal
urllib3
vine
wcwidth
Werkzeug
WTForms